refactor(common): update logger
