+++ packages/common/services/interfaces/index.ts
@@ -1,0 +1,1 @@
+export * from './console-logger-options.interface';
