+++ packages/common/services/utils/is-log-level-enabled.util.ts
@@ -1,8 +1,8 @@
 import { LogLevel } from '../logger.service';
 
 const LOG_LEVEL_VALUES: Record<LogLevel, number> = {
-  debug: 0,
-  verbose: 1,
+  verbose: 0,
+  debug: 1,
   log: 2,
   warn: 3,
   error: 4,
